make the prelude an actual prelude
